+++ truck-js/tests/test.js
@@ -9,13 +9,13 @@ Deno.test("tseep cube", async () => {
   const abst = Truck.tsweep(f, [0.0, 0.0, 1.0]);
   const solid = abst.into_solid();
   const vec = solid.to_json();
-  const readVec = await Deno.readFile("./tests/cube.json");
+  const readVec = await Deno.readFile("../resources/shape/cube.json");
   assertEquals(vec, readVec);
 });
 
 Deno.test("rseep torus", async () => {
   await init(Deno.readFile("./pkg/truck_js_bg.wasm"));
-  const v = Truck.vertex(0.5, 0.0, 0.0);
+  const v = Truck.vertex(0.75, 0.0, 0.25);
   const w = Truck.rsweep(
     v.upcast(),
     [0.75, 0.0, 0.0],
@@ -31,6 +31,6 @@ Deno.test("rseep torus", async () => {
   const shell = abst.into_shell();
   const solid = shell.into_solid();
   const vec = solid.to_json();
-  const readVec = await Deno.readFile("./tests/torus.json");
+  const readVec = await Deno.readFile("../resources/shape/torus.json");
   assertEquals(vec, readVec);
 });